Update the roster outside of the volume indicator update.
This change moves the work of adding attendees to the roster, and
fetching and setting their name, to the presence change handler. This
avoids us doing more work when the volume indicator changes, and also
ensures that attendees are added to the roster before the audio callback
is invoked.

This commit also shifts some logic out into named functions for clarity
and efficiency.